feat: ajustes do modelo de boletim
